Support remotely-opened RTCDataChannel
RTCDataChannel may be opened:

- either locally (which is already supported in react-native-webrtc)
using the Peer-to-peer Data API which extends the RTCPeerConnection
interface with the method createDataChannel

- or remotely (i.e. by the other peer). The local peer receives a
datachannel event through RTCPeerConnection. That's defined by the same
Peer-to-peer Data API.

In order to add support for the second scenario, remove reactTag from
DataChannel and use the streamId property of the Objective-C
RTCDataChannel as the id of the JavaScript counterpart (as done by the
browser API).