Fix the delivery of RTCDataChannel text messages
1. NSString's stringFromUTF8String expects NSData's bytes to be
terminated by a null (character). Which is not always the case with
NSData because it has a length property and, thus, terminating by a null
(character) is not mandatory. In such a case, RTCDataChannel would fail
to properly convert NSData to an NSString and would instead convert it
to nil.

2. In the above scenario, RTCDataChannel would crash the process because
it would attempt to insert a nil into an NSDictionary. NSDictionary does
not allow nil keys and/or values.