Add ability to switch cameras on the fly
This patch adds a private custom API for switching the cameras on a
video track on the fly, without the need for adding / removing tracks
or renegotiating.

It's MediaStreamTrack._switchCamera() and it only works for local video
streams, it will throw otherwise.

This is analogous to how the reference application (AppRTCMobile) implements
camera switching.