android: fix transparent rendering if no frames are received
A SurfaceView is composed of a View and a Surface which "punches a hole"
through it.

First set the view background to black so the surface doesn't punch
through it. The in the callback for the first received frame, set the
background to transparent, thus allowing the surface (which now has
actual video content) to be seen.