android: avoid crashes when adding / removing renderers
Stream operations run in the WebRTC thread, but the view methids run in
the UI thread. So there is a window of opportunity for Bad Things (TM)
to happen when a stream is released (in the WebRTC thread).

This bug has existed for a while, but somewhere between M69 and M75
WebRTC added some more spartan checks to the MediaStream operations, so
they raise IllegalStateException when the stream was released. Let's
ignore those.